# Api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Voting service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/auth/reset-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Request password reset for a registered user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password",
                "parameters": [
                    {
                        "description": "User email for password reset",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/__.ResetPasswordReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.Empty"
                        }
                    },
                    "400": {
                        "description": "Error while requesting password reset",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while requesting password reset",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/progress/achievements/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the achievements of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Progress"
                ],
                "summary": "Get user achievements",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.AchievementsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/progress/by_language/{language_id}/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the progress of a user for a specific language and user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Progress"
                ],
                "summary": "Get user progress by language and user ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language ID",
                        "name": "language_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.ProgressByLanguageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/progress/daily/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the daily progress of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Progress"
                ],
                "summary": "Get daily progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.DailyProgressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/progress/leader/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the leaderboard for a specific language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Progress"
                ],
                "summary": "Get leaderboard by language",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language ID",
                        "name": "language_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.LeaderboardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/progress/monthly/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the monthly progress of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Progress"
                ],
                "summary": "Get monthly progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.MonthlyProgressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/progress/skills/{language_id}/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the skills of a user for a specific language",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Progress"
                ],
                "summary": "Get user skills by language",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Language ID",
                        "name": "language_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.SkillsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/progress/weekly/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the weekly progress of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Progress"
                ],
                "summary": "Get weekly progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/__.WeeklyProgressResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "__.Achievement": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "earned_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "__.AchievementsResponse": {
            "type": "object",
            "properties": {
                "achievements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/__.Achievement"
                    }
                }
            }
        },
...
        "__.DailyProgressResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "lessons_completed": {
                    "type": "integer"
                },
                "minutes_practiced": {
                    "type": "integer"
                },
                "new_words_learned": {
                    "type": "integer"
                },
                "streak_days": {
                    "type": "integer"
                },
                "xp_earned": {
                    "type": "integer"
                }
            }
        },
        "__.Empty": {
            "type": "object",
            "properties": {
                "isTrue": {
                    "type": "boolean"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "__.LeaderboardEntry": {
            "type": "object",
            "properties": {
                "level": {
                    "type": "string"
                },
                "rank": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                },
                "xp": {
                    "type": "integer"
                }
            }
        },
        "__.LeaderboardResponse": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string"
                },
                "leaderboard": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/__.LeaderboardEntry"
                    }
                },
                "user_rank": {
                    "type": "integer"
                }
            }
        },
...
        "__.MonthlyProgressResponse": {
            "type": "object",
            "properties": {
                "lessons_completed": {
                    "type": "integer"
                },
                "month": {
                    "type": "string"
                },
                "most_improved_skill": {
                    "type": "string"
                },
                "new_words_learned": {
                    "type": "integer"
                },
                "total_minutes_practiced": {
                    "type": "integer"
                },
                "total_xp_earned": {
                    "type": "integer"
                }
            }
        },
...
        "__.ProgressByLanguageResponse": {
            "type": "object",
            "properties": {
                "grammar_rules_mastered": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "lessons_completed": {
                    "type": "integer"
                },
                "level": {
                    "type": "string"
                },
                "listening_comprehension": {
                    "type": "integer"
                },
                "speaking_fluency": {
                    "type": "integer"
                },
                "vocabulary_learned": {
                    "type": "integer"
                },
                "xp": {
                    "type": "integer"
                }
            }
        },
...
        "__.ResetPasswordReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "__.Skill": {
            "type": "object",
            "properties": {
                "comprehension_rate": {
                    "type": "integer"
                },
                "fluency_rate": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "progress_to_next_level": {
                    "type": "integer"
                },
                "rules_mastered": {
                    "type": "integer"
                },
                "words_learned": {
                    "type": "integer"
                }
            }
        },
        "__.SkillsResponse": {
            "type": "object",
            "properties": {
                "language": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/__.Skill"
                    }
                }
            }
        },
...
        "__.WeeklyProgressResponse": {
            "type": "object",
            "properties": {
                "lessons_completed": {
                    "type": "integer"
                },
                "most_active_day": {
                    "type": "string"
                },
                "new_words_learned": {
                    "type": "integer"
                },
                "total_minutes_practiced": {
                    "type": "integer"
                },
                "total_xp_earned": {
                    "type": "integer"
                },
                "week_end": {
                    "type": "string"
                },
                "week_start": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authourization",
            "in": "header"
        }
    }
}